backend/requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
openai==1.3.0
httpx==0.24.0
python-dotenv==0.19.0
python-multipart==0.0.5
aiofiles==0.7.0
pydantic>=1.10.13,<2